+++ Changes.md
@@ -6,6 +6,7 @@ Version 3.3.0.3
      'Section - Section that contains the earliest "Not attempted" activity'.
   3. Fixed cosmetic issues with Framed sections.  Note:  The frame around the section will change when editing to
      accomodate the elements within the sides.
+  4. Implemented 'Edit mode' -> 'Add section below'.
 
 Version 3.3.0.2
   1. Added progress tooltip options.